pyproject.toml
[project]
name = "message-stream-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.4.0",
    "asyncio>=3.4.3",
    "datasets>=3.3.2",
    "ollama>=0.4.7",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "websockets>=15.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
]

[project.scripts]
ingest = "client:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]